fuse Try thread schedule first before going to sleep
There might be work immediately after going to sleep and that
wakeup might have a high latency.

With this changes CPU time of passthrough_hp roughly doubles,
but performance numbers do as well

Without the patch
1.98,1.98,imesrv1,1,1648844145,,,,,,,,,,,,,,,,,30,1,1,,10,2183,30,3318,42,4242,24,2292,31,2469,41,3729,28,,,,,,,50747us,1011us,2254us,4678us,820us,44645us

With the patch
1.98,1.98,imesrv1,1,1648852632,,,,,,,,,,,,,,,,,30,1,1,,10,4809,34,7770,52,8418,27,4901,35,5451,51,5768,31,,,,,,,40273us,559us,1801us,385us,512us,263ms

Note: The numbers are very cpufreq governor dependent, this is with
the "conservative" scheduler, which seems to be the ubuntu
default.


FIXME: Add performance gov values.


This is a from a single threaded bonnie++
bonnie++ -q -s0 -d <path> -n 30:1:1:10 -r 0