Allow non-extending parallel direct writes on the same file.
In general, as of now, in FUSE, direct writes on the same file are
serialized over inode lock i.e we hold inode lock for the full duration
of the write request. I could not find in fuse code and git history
a comment which clearly explains why this exclusive lock is taken
for direct writes.  Following might be the reasons for acquiring
an exclusive lock but not be limited to
1) Our guess is some USER space fuse implementations might be relying
   on this lock for seralization.
2) The lock protects against file read/write size races.
3) Ruling out any issues arising from partial write failures.

This patch relaxes the exclusive lock for direct non-extending writes
only. File size extending writes might not need the lock either,
but we are not entirely sure if there is a risk to introduce any
kind of regression. Furthermore, benchmarking with fio does not
show a difference between patch versions that take on file size
extension a) an exclusive lock and b) a shared lock.

A possible example of an issue with i_size extending writes are write
error cases. Some writes might succeed and others might fail for
file system internal reasons - for example ENOSPACE. With parallel
file size extending writes it _might_ be difficult to revert the action
of the failing write, especially to restore the right i_size.

With these changes, we allow non-extending parallel direct writes
on the same file with the help of a flag called
FOPEN_PARALLEL_DIRECT_WRITES. If this flag is set on the file (flag is
passed from libfuse to fuse kernel as part of file open/create),
we do not take exclusive lock anymore, but instead use a shared lock
that allows non-extending writes to run in parallel.
FUSE implementations which rely on this inode lock for serialisation
can continue to do so and serialized direct writes are still the
default.  Implementations that do not do write serialization need to
be updated and need to set the FOPEN_PARALLEL_DIRECT_WRITES flag in
their file open/create reply.

On patch review there were concerns that network file systems (or
vfs multiple mounts of the same file system) might have issues with
parallel writes. We believe this is not the case, as this is just a
local lock, which network file systems could not rely on anyway.
I.e. this lock is just for local consistency.

Benchmark Results for single shared file writes (SSF)

1) MPIIO - extending parallel writes only possible with some
tools and if so, only as manual operation

Results with https://github.com/dchirikov/mpiio.git

        unpatched    patched            patched
        (extending) (extending)         (manual extending)
----------------------------------------------------------
              MB/s         MB/s          MB/s
2 threads     2275.00      2497.00      5688.00
4 threads     2438.00      2560.00      10240.00
8 threads     2925.00      3792.00      25600.00
16 threads    3792.00     10240.00      20480.00

(Patched-nonextending is a manual operation on the file to extend the size, mpiio does not support that natively, as far as I know.)

Results with IOR (HPC quasi standard benchmark)

ior -w -E -k -o /tmp/test/home/hbi/test/test.1 -a mpiio -s 1280 -b 8m -t 8m

        unpatched         patched
        (extending)     (extending)    gain
--------------------------------------------
             MB/s           MB/s       %
2 threads    2086.10        2027.76   -2.8
4 threads    1858.94        2132.73   +14.7
8 threads    1792.68        4609.05   +157.1
16 threads   1786.48        8627.96   +383.0

2) fio single shared file

Job file for SSF:
[global]
directory=/tmp/dest
filename=ssf
size=100g
blocksize=1m
ioengine=sync
group_reporting=1
fallocate=none
runtime=60
stonewall

[write]
rw=randwrite:256
rw_sequencer=sequential
fsync_on_close=1

Job file for file-per-job:
[sequential-write]
rw=write
size=100G
directory=/tmp/dest/
group_reporting
name=sequential-write-direct
bs=1M
runtime=60

        unpatched         patched    gain
-------------------------------------------
              MiB/s        MiB/s       %
1 thread      3624         3137       -13.4
2 threads     5801         7736       +33.4
4 threads     4794         14400      +200.4
8 threads     3946         22600      +472.7
16 threads    4040         25600      +533.7
32 threads    2822         20200      +615.8

Signed-off-by: Dharmendra Singh <dsingh@ddn.com>
Signed-off-by: Bernd Schubert <bschubert@ddn.com>
Tested-by: Horst Birthelmer <hbirthelmer@ddn.com>