fuse: Mark blocking dio and use for uring
blocking requests are better handled on the same core as
the current core is probably not doing anything while
the request is handled by the ring/daemon.
So far ring-async was entirely related to FR_BACKGROUND,
but that flag is also set for blocking direct-io
requests. A new bit is added into struct fuse_io_args
and checked by the function that makes the uring queue
decision.

Signed-off-by: Bernd Schubert <bschubert@ddn.com>