Date:   Tue,  7 Mar 2023 23:31:57 -0800 Subject: [PATCH 2/6] sched: add WF_CURRENT_CPU and externise ttwu From: Peter Oskolkov <posk@google.com>
Add WF_CURRENT_CPU wake flag that advices the scheduler to
move the wakee to the current CPU. This is useful for fast on-CPU
context switching use cases.

In addition, make ttwu external rather than static so that
the flag could be passed to it from outside of sched/core.c.

Signed-off-by: Peter Oskolkov <posk@google.com>
Signed-off-by: Andrei Vagin <avagin@google.com>