fuse over-io-uring: Avoid hanging umount after daemon termination
umount (not 'umount -f') was hanging on umount when the
server/daemon was already terminated. Reason is that
waiting io_uring commands have a /dev/fuse file reference and with
that ->release is not called and the connection is then not
terminated. Umount then tries to send a getattr request which
is not completed until manual connection abort through sysfs.

Solution here is to enforce a clone of the /dev/fuse file
for every queue and ->release() of the main file (basically control)
file then aborts ring queues and terminates the connection.